refactor(xmpp): Promote MUC mod to directory
